backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.1
rdkit-pypi==2022.9.5
networkx==3.2.1
pydantic==2.6.1
requests==2.31.0
openai>=1.40.0
numpy<2.0.0
